Fix for issues caused by merges

+++ common/src/desktopMain/kotlin/chat/sphinx/common/components/notifications/DesktopSphinxNotificationManager.kt
@@ -8,7 +8,7 @@ object DesktopSphinxNotificationManager: SphinxNotificationManager {
     val notifications = mutableStateListOf<Pair<String, String>>()
 
     override fun notify(
-        notificationId: Int,
+        notificationId: Long,
         groupId: String?,
         title: String,
         message: String
